# docs/swagger.yaml
basePath: /api/v1
definitions:
  audit.AuditLogEvent:
    properties:
      action:
        type: string
      actorEmail:
        type: string
      actorUserID:
        format: int64
        type: integer
      metadata:
        additionalProperties: {}
        type: object
      reason:
        type: string
      resource:
        type: string
      success:
        type: boolean
      targetEmail:
        type: string
      targetUserID:
        format: int64
        type: integer
      timeStamp:
        format: int64
        type: integer
    type: object
  audit.AuditLogResponse:
    properties:
      events:
        items:
          $ref: '#/definitions/audit.AuditLogEvent'
        type: array
      next_cursor:
        type: integer
    type: object
...
  ldap.CreateLdapUserRequest:
    properties:
      email:
        type: string
    type: object
  ldap.LdapConfigRequest:
    properties:
      base_dn:
        type: string
      bind_dn:
        type: string
      bind_password:
        type: string
      host:
        type: string
      port:
        type: integer
      skip_tls_verify:
        type: boolean
      use_ssl:
        type: boolean
    type: object
  ldap.LdapConfigResponse:
    properties:
      base_dn:
        type: string
      bind_dn:
        type: string
      host:
        type: string
      port:
        type: integer
      skip_tls_verify:
        type: boolean
      use_ssl:
        type: boolean
    type: object
  ldap.LdapSearchRequest:
    properties:
      query:
        type: string
    type: object
  ldap.LdapSearchResponse:
    properties:
      display_name:
        type: string
      email:
        type: string
    type: object
...
info:
  contact: {}
  description: 'All endpoints return responses wrapped in a standard envelope: {"success":
    bool, "data": (schema shown per endpoint), "error": string}. All protected endpoints
    can return 401.'
  title: ez2boot API
  version: "1.0"
paths:
...
  /audit/events:
    get:
      description: '[Admin] retrieve audit events'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/audit.AuditLogResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Get audit events
      tags:
      - audit
  /auth/ldap:
    delete:
      description: '[Admin] delete an LDAP config. No payload or selector, as there
        can be only one configuration.'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
      security:
      - BasicAuth: []
      summary: Delete LDAP config
      tags:
      - ldap
    get:
      description: '[Admin] get the current LDAP config'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ldap.LdapConfigResponse'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Get LDAP config
      tags:
      - ldap
    post:
      consumes:
      - application/json
      description: '[Admin] set an LDAP config'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ldap.LdapConfigRequest'
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Set LDAP config
      tags:
      - ldap
  /auth/ldap/users/search:
    post:
      consumes:
      - application/json
      description: '[Admin] search the LDAP service for a user during provisioning.
        Matches where UPN starts with query and returns one match only.'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ldap.LdapSearchRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ldap.LdapSearchResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "503":
          description: Service Unavailable
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Search for LDAP user
      tags:
      - ldap
...
  /user/ldap:
    post:
      consumes:
      - application/json
      description: '[Admin] provision an LDAP user. Assumes an LDAP config is present
        and the user exists in the IDP.'
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ldap.CreateLdapUserRequest'
      responses:
        "201":
          description: Created
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "409":
          description: Conflict
          schema:
            type: string
      security:
      - BasicAuth: []
      summary: Create LDAP user
      tags:
      - ldap
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"